+++ lighthouserc.js
@@ -1,0 +1,19 @@
+module.exports = {
+  ci: {
+    collect: {
+      url: [
+        '/',
+        '/404.html',
+        '/resume/',
+        '/releases/',
+        '/tools/',
+        '/portfolio/',
+      ],
+      staticDistDir: './dist',
+      numberOfRuns: 5,
+    },
+    upload: {
+      target: 'temporary-public-storage',
+    },
+  },
+};