+++ test/e2e/portfolio_test.js
@@ -1,0 +1,13 @@
+/* eslint-disable no-undef */
+
+const name = 'portfolio';
+
+Feature(name);
+
+Scenario('Heading 1', async (I) => {
+  I.amOnPage(`/${name}`);
+  I.see('PORTFOLIO', 'h1');
+
+  I.saveScreenshot(`${name}.png`);
+  await I.seeVisualDiff(`${name}.png`);
+});